stripe2aws/requirements.txt
boto3==1.40.6
botocore==1.40.6
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
urllib3==1.26.20
